requirements.txt
aiofiles==24.1.0
albucore==0.0.24
albumentations==2.0.8
git+https://github.com/Alexandre-Delplanque/HerdNet.git@7e25f482d875522c59c446dc0c78c8f6f2dd448d#egg=animaloc
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
brotli==1.2.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fastapi==0.121.0
ffmpy==0.6.4
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
gradio==5.49.1
gradio_client==1.13.3
groovy==0.1.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.2
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python-headless==4.12.0.88
orjson==3.11.4
packaging==25.0
pandas==2.3.3
pillow==11.3.0
platformdirs==4.5.0
protobuf==6.33.0
pydantic==2.11.10
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
ruff==0.14.4
safehttpx==0.1.7
scikit-learn==1.7.2
scipy==1.16.3
semantic-version==2.10.0
sentry-sdk==2.43.0
setuptools==80.9.0
shellingham==1.5.4
simsimd==6.5.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.49.3
stringzilla==4.2.3
sympy==1.14.0
threadpoolctl==3.6.0
tomlkit==0.13.3
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
wandb==0.22.3
websockets==15.0.1
